requirements.txt
python-telegram-bot==21.6
discord.py==2.4.0
aiohttp==3.10.5
python-dotenv==1.0.1
